Refactor ChangelogImpl
