+++ packages/sfpowerscripts-cli/src/commands/sfpowerscripts/changelog/generate.ts
@@ -1,14 +1,7 @@
 import { flags, SfdxCommand } from '@salesforce/command';
 import { Messages } from '@salesforce/core';
-import ArtifactFilePathFetcher, { ArtifactFilePaths } from "@dxatscale/sfpowerscripts.core/lib/artifacts/ArtifactFilePathFetcher";
-import PackageMetadata from "@dxatscale/sfpowerscripts.core/lib/PackageMetadata";
-import { ReleaseChangelog, Release, Artifact } from "@dxatscale/sfpowerscripts.core/lib/changelog/interfaces/ReleaseChangelogInterfaces";
-import { Changelog as PackageChangelog } from "@dxatscale/sfpowerscripts.core/lib/changelog/interfaces/GenericChangelogInterfaces";
-import generateMarkdown from "@dxatscale/sfpowerscripts.core/lib/changelog/GenerateChangelogMarkdown";
-import * as fs from "fs-extra"
-import path = require('path');
-import simplegit, { SimpleGit } from "simple-git/promise";
-const tmp = require('tmp');
+import ChangelogImpl from '../../../impl/changelog/ChangelogImpl';
+
 
 
 Messages.importMessagesDirectory(__dirname);
@@ -70,199 +63,21 @@ export default class GenerateChangelog extends SfdxCommand {
     })
   };
 
-
-  public async run(){
-    let tempDir = tmp.dirSync({unsafeCleanup: true});
+  async run() {
     try {
-      let git: SimpleGit = simplegit();
-
-      console.log(`Cloning repository ${this.flags.repourl}`);
-      await git.clone(
+      let changelogImpl: ChangelogImpl = new ChangelogImpl(
+        this.flags.artifactdir,
+        this.flags.releasename,
+        this.flags.workitemfilter,
         this.flags.repourl,
-        tempDir.name
-      );
-      const repoTempDir = tempDir.name;
-
-      console.log(`Checking out branch ${this.flags.branchname}`);
-      git = simplegit(repoTempDir);
-      await git.checkout(this.flags.branchname);
-
-      let artifact_filepaths: ArtifactFilePaths[] = ArtifactFilePathFetcher.fetchArtifactFilePaths(
-        this.flags.artifactdir
-      );
-
-      if (artifact_filepaths.length === 0) {
-        throw new Error(`No artifacts found at ${path.resolve(process.cwd(), this.flags.artifactdir)}`);
-      }
-
-      let prevReleaseDefinition: Release;
-      let releaseChangelog: ReleaseChangelog;
-      if (fs.existsSync(path.join(repoTempDir,`releasechangelog.json`))) {
-        releaseChangelog = JSON.parse(fs.readFileSync(path.join(repoTempDir,`releasechangelog.json`), 'utf8'));
-        if (releaseChangelog["releases"].length > 0) {
-          prevReleaseDefinition = releaseChangelog["releases"][releaseChangelog["releases"].length - 1];
-        }
-      }
-
-      if (prevReleaseDefinition?.name === this.flags.releasename) {
-        console.log(`The release named "${this.flags.releasename}" already exists.`);
-        console.log("Skipping changelog generation...");
-        return;
-      }
-
-      let packageChangelogMap: {[P:string]: string} = {};
-      let latestReleaseDefinition: Release = {
-        name: this.flags.releasename,
-        workItems: {},
-        artifacts: []
-      };
-
-      // Read artifacts for latest release definition
-      let missingChangelogs: Error[] = [];
-      for (let artifactFilepaths of artifact_filepaths ) {
-
-        let packageMetadata: PackageMetadata = JSON.parse(
-          fs.readFileSync(artifactFilepaths.packageMetadataFilePath, 'utf8')
-        );
-
-        let artifact: Artifact = {
-          name: packageMetadata["package_name"],
-          from: undefined,
-          to: packageMetadata["sourceVersion"]?.slice(0,8) || packageMetadata["sourceVersionTo"]?.slice(0,8),
-          version: packageMetadata["package_version_number"],
-          latestCommitId: undefined,
-          commits: undefined
-        }
-
-        latestReleaseDefinition["artifacts"].push(artifact);
-
-        if (!fs.existsSync(artifactFilepaths.changelogFilePath)) {
-          missingChangelogs.push(
-            new Error(`No changelog found in artifact ${packageMetadata["package_name"]} ${packageMetadata["package_version_number"]}`)
-          );
-        }
-
-        packageChangelogMap[packageMetadata["package_name"]] = artifactFilepaths.changelogFilePath;
-      }
-
-      if (missingChangelogs.length > 0) {
-        throw missingChangelogs;
-      }
-
-      console.log("Generating changelog...");
-
-      let prevReleaseLatestCommitId: {[P: string]: string} = {}
-      if (prevReleaseDefinition) {
-        for (let artifact of latestReleaseDefinition["artifacts"]) {
-          for (let prevReleaseArtifact of prevReleaseDefinition["artifacts"]) {
-            if (artifact["name"] === prevReleaseArtifact["name"]) {
-              artifact["from"] = prevReleaseArtifact["to"];
-              prevReleaseLatestCommitId[artifact["name"]] = prevReleaseArtifact["latestCommitId"];
-              break;
-            }
-          }
-        }
-      }
-
-      // Get commits for the latest release
-      for (let artifact of latestReleaseDefinition["artifacts"]) {
-        let packageChangelog: PackageChangelog = JSON.parse(
-          fs.readFileSync(packageChangelogMap[artifact["name"]], 'utf8')
-        );
-
-        artifact["latestCommitId"] = packageChangelog["commits"][0]["commitId"];
-
-        let fromIdx;
-        if (artifact["from"]) {
-          fromIdx = packageChangelog["commits"].findIndex( (commit) =>
-            commit["commitId"] === prevReleaseLatestCommitId[artifact["name"]]
-          );
-          if (fromIdx === -1) {
-            console.log(`Cannot find commit Id ${prevReleaseLatestCommitId[artifact["name"]]} in ${artifact["name"]} changelog`);
-            console.log("Assuming that there are no changes...");
-            artifact["commits"] = [];
-            continue;
-          }
-        }
-
-
-        if (fromIdx > 0) {
-          artifact["commits"] = packageChangelog["commits"].slice(0, fromIdx);
-        } else if (fromIdx === 0) {
-          // Artifact verison has not changed
-          artifact["commits"] = [];
-          // Skip to next artifact
-          continue;
-        } else if (fromIdx === undefined ) {
-          // Artifact was not in previous release
-          artifact["commits"] = packageChangelog["commits"];
-        }
-
-
-        // Compute work items for latest release
-        let workItemFilter: RegExp = RegExp(this.flags.workitemfilter, 'gi');
-        for (let commit of artifact["commits"]) {
-          let commitMessage: String = commit["message"] + "\n" + commit["body"];
-          let workItems: RegExpMatchArray = commitMessage.match(workItemFilter);
-          if (workItems) {
-              for (let item of workItems) {
-                  if (latestReleaseDefinition["workItems"][item] == null) {
-                      latestReleaseDefinition["workItems"][item] = new Set<string>();
-                      latestReleaseDefinition["workItems"][item].add(commit["commitId"].slice(0,8));
-                  } else {
-                      latestReleaseDefinition["workItems"][item].add(commit["commitId"].slice(0,8));
-                  }
-              }
-          }
-        }
-      }
-
-      // Convert each work item Set to Array
-      // Enables JSON stringification of work item
-      for (let key in latestReleaseDefinition["workItems"]) {
-        latestReleaseDefinition["workItems"][key] = Array.from(latestReleaseDefinition["workItems"][key]);
-      }
-
-      // Append results to release changelog
-      if (releaseChangelog) {
-        releaseChangelog["releases"].push(latestReleaseDefinition);
-      } else {
-        releaseChangelog = {
-          releases: [latestReleaseDefinition]
-        }
-      }
-
-      fs.writeFileSync(
-        path.join(repoTempDir,`releasechangelog.json`),
-        JSON.stringify(releaseChangelog, null, 4)
-      );
-
-      let payload: string = generateMarkdown(
-        releaseChangelog,
-        this.flags.workitemurl,
+        this.flags.branchname,
         this.flags.limit,
-        this.flags.showallartifacts
+        this.flags.workitemurl,
+        this.flags.showallartifacts,
+        this.flags.forcepush
       );
 
-      fs.writeFileSync(
-        path.join(repoTempDir,`Release-Changelog.md`),
-        payload
-      );
-
-      console.log("Pushing changelog files to", this.flags.repourl, this.flags.branchname);
-      git = simplegit(repoTempDir);
-      await git.addConfig("user.name", "sfpowerscripts");
-      await git.addConfig("user.email", "sfpowerscripts@dxscale");
-      await git.add([`releasechangelog.json`, `Release-Changelog.md`]);
-      await git.commit(`[skip ci] Updated Changelog ${this.flags.releasename}`);
-
-      if (this.flags.forcepush) {
-        await git.push(`--force`);
-      } else {
-        await git.push();
-      }
-
-      console.log(`Successfully generated changelog`);
+      await changelogImpl.exec();
     } catch (err) {
 
       let errorMessage: string = "";
@@ -275,11 +90,7 @@ export default class GenerateChangelog extends SfdxCommand {
       }
       console.log(errorMessage);
 
-      tempDir.removeCallback();
-
       process.exit(1);
-    } finally {
-      tempDir.removeCallback();
     }
   }
 }
